173 Change setSpanAttribute to setSpanAttributes and revert withSpanAttribute
This ensures backwards compatibility of existing pipelines.
