+++ src/test/java/io/jenkins/plugins/opentelemetry/job/step/SetSpanAttributesStepTest.java
@@ -18,30 +18,28 @@ import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.junit.Test;
 
-import java.util.Optional;
-
 import static org.junit.Assume.assumeFalse;
 
-public class SetSpanAttributeStepTest extends BaseIntegrationTest {
+public class SetSpanAttributesStepTest extends BaseIntegrationTest {
 
     @Test
-    public void testSimplePipelineWithSetSpanAttributeStepTarget() throws Exception {
+    public void testSimplePipelineWithSetSpanAttributesStep() throws Exception {
         assumeFalse(SystemUtils.IS_OS_WINDOWS);
         // BEFORE
 
         String pipelineScript = "def xsh(cmd) {if (isUnix()) {sh cmd} else {bat cmd}};\n" +
-            "setSpanAttribute(key: 'pipeline.type', value: 'release', target: 'PIPELINE_ROOT_SPAN')\n" +
+            "setSpanAttributes([spanAttribute(key: 'pipeline.type', value: 'release', target: 'PIPELINE_ROOT_SPAN')])\n" +
             "node() {\n" +
             "    stage('build') {\n" +
-            "       setSpanAttribute(key: 'pipeline.importance', value: 'critical', target: 'PIPELINE_ROOT_SPAN')\n" +
-            "       setSpanAttribute(key: 'stage.type', value: 'build-java-maven', target: 'CURRENT_SPAN')\n" +
-            "       setSpanAttribute(key: 'build.tool', value: 'maven')\n" +
+            "       setSpanAttributes([spanAttribute(key: 'pipeline.importance', value: 'critical', target: 'PIPELINE_ROOT_SPAN')])\n" +
+            "       setSpanAttributes([spanAttribute(key: 'stage.type', value: 'build-java-maven', target: 'CURRENT_SPAN')])\n" +
+            "       setSpanAttributes([spanAttribute(key: 'build.tool', value: 'maven'), spanAttribute(key: 'test.tool', value: 'junit')])\n" +
             "       xsh (label: 'release-script', script: 'echo ze-echo-1') \n" +
             "    }\n" +
             "}";
         jenkinsRule.createOnlineSlave();
 
-        final String jobName = "test-simple-pipeline-with-with-span-attribute-step-target" + jobNameSuffix.incrementAndGet();
+        final String jobName = "test-simple-pipeline-with-with-span-attribute-step" + jobNameSuffix.incrementAndGet();
         WorkflowJob pipeline = jenkinsRule.createProject(WorkflowJob.class, jobName);
         pipeline.setDefinition(new CpsFlowDefinition(pipelineScript, true));
         jenkinsRule.assertBuildStatus(Result.SUCCESS, pipeline.scheduleBuild2(0));
@@ -77,6 +75,12 @@ public class SetSpanAttributeStepTest extends BaseIntegrationTest {
             SpanData actualSpanData = spans.breadthFirstStream().filter(sdw -> "Stage: build".equals(sdw.spanData.getName())).findFirst().get().spanData;
             String actualBuildTool = actualSpanData.getAttributes().get(AttributeKey.stringKey("build.tool"));
             MatcherAssert.assertThat(actualBuildTool, CoreMatchers.is("maven"));
+        }
+
+        { // attribute 'test.tool' - implicitly CURRENT_SPAN, multiple spanAttributes specified in list
+            SpanData actualSpanData = spans.breadthFirstStream().filter(sdw -> "Stage: build".equals(sdw.spanData.getName())).findFirst().get().spanData;
+            String actualBuildTool = actualSpanData.getAttributes().get(AttributeKey.stringKey("test.tool"));
+            MatcherAssert.assertThat(actualBuildTool, CoreMatchers.is("junit"));
         }
 
         MatcherAssert.assertThat(spans.cardinality(), CoreMatchers.is(8L));
